Default to Main_Zone when sending commands to receiver
Previously, we would send a request with no zone (which is invalid XML) if the user has not set the zone in the app. Issue found by da_man on SmartThings Community: https://community.smartthings.com/t/release-yamaha-receiver-on-off/228789/7

Tested by editing preferences in the IDE and setting zone to "". This is not quite the same as unset preference, so I'm not 100% confident this is correct. I will as da_man to test to verify.