Fix setVolume
It seems that a volume option in the app has appeared, so I can now test setting Volume. this has uncovered two issues fixed in this commit:

* The XML command for setting volume was incorrect, now fixed.
* The calculation of volume to set was missing a factor of 10, now fixed.

There still seems to be an issue showing the volume in the UI (perhaps I'm setting the wrong attribute) but setting a volume does work.